Add directory emp

+++ frm101.pas
@@ -21,22 +21,32 @@ type
     btnAdd: TButton;
     bntUpdate: TButton;
     btnDelete: TButton;
-    Panel3: TPanel;
-    Label2: TLabel;
-    Button1: TButton;
-    Edit2: TEdit;
     Panel4: TPanel;
     Label3: TLabel;
     btnSearchDept: TButton;
     Edit3: TEdit;
     DBGrid2: TDBGrid;
     DBGrid3: TDBGrid;
+    Panel3: TPanel;
+    Label2: TLabel;
+    btnSearchEmp: TButton;
+    Edit2: TEdit;
+    btnResetPost: TButton;
+    btnResetEmp: TButton;
+    btnResetDept: TButton;
     procedure btnAddClick(Sender: TObject);
     procedure btnSearchPostClick(Sender: TObject);
     procedure btnDeleteClick(Sender: TObject);
     procedure bntUpdateClick(Sender: TObject);
     procedure FormCreate(Sender: TObject);
     procedure btnSearchDeptClick(Sender: TObject);
+    procedure TabSheet1Show(Sender: TObject);
+    procedure TabSheet3Show(Sender: TObject);
+    procedure TabSheet2Show(Sender: TObject);
+    procedure btnSearchEmpClick(Sender: TObject);
+    procedure btnResetPostClick(Sender: TObject);
+    procedure btnResetEmpClick(Sender: TObject);
+    procedure btnResetDeptClick(Sender: TObject);
   private
     { Private declarations }
   public
@@ -48,7 +58,7 @@ var
 
 implementation
 
-uses DM2, frm104, frm105;
+uses DM2, frm104, frm105, frm106;
 
 {$R *.dfm}
 
@@ -57,6 +67,7 @@ begin
 // ***** Справочник - Должности  *****
   if Pagecontrol1.TabIndex = 0 then
     begin
+    //  Form104 := TForm104.Create(Self);
       Form104.btnAdd.Visible := True;
       Form104.btnUpdate.Visible := False;
       Form104.Edit1.Clear;
@@ -66,6 +77,7 @@ begin
 // ***** Справочник - Отделы  *****
   if Pagecontrol1.TabIndex = 1 then
      begin
+   //   Form105 := TForm105.Create(Self);
       Form105.btnAdd.Visible := True;
       Form105.btnUpdate.Visible := False;
       Form105.Edit1.Clear;
@@ -74,7 +86,16 @@ begin
     end;
 // ***** Справочник - Сотрудники  *****
   if Pagecontrol1.TabIndex = 2 then
-    showmessage('2');
+     begin
+    //  Form106 := TForm106.Create(Self);
+      Form106.btnAdd.Visible := True;
+      Form106.btnUpdate.Visible := False;
+      Form106.Edit1.Clear;
+      Form106.MaskEdit1.Clear;
+      Form106.Combobox1.Text := '';
+      Form106.Combobox2.Text := '';
+      Form106.Show;
+     end;
 end;
 
 procedure TForm101.btnSearchPostClick(Sender: TObject);
@@ -93,37 +114,38 @@ end;
 
 procedure TForm101.btnDeleteClick(Sender: TObject);
 var
-  v_id: String;
+  v_id, v_full_name: String;
+  v_fired_flag: Integer;
 begin
 // ***** Справочник - Должности  *****
- if MessageDlg('Вы точно хотите удалить строку',mtConfirmation, [mbYes, mbNo], 0) = mrYes then
-   begin
      if Pagecontrol1.TabIndex = 0 then
-      begin
-       if DBGrid1.DataSource.DataSet.RecordCount = 0 then
-         ShowMessage('Ничего не выбрано');
-       if DBGrid1.DataSource.DataSet.RecordCount > 0 then
-        begin
-          v_id := DBGrid1.DataSource.DataSet.FieldByName('id').AsString;
-          with DM_main.QryEmpty do
-            begin
-              SQL.Clear;
-              SQL.Add('DELETE FROM positions ');
-              SQL.Add(' WHERE id = ' + v_id + ' ');
-               try ExecSQL except
-                on E: EDatabaseError do
-                 ShowMessage (E.Message);
+       begin
+         if DBGrid1.DataSource.DataSet.RecordCount <= 0 then
+           ShowMessage('Ничего не выбрано');
+         if (DBGrid1.DataSource.DataSet.RecordCount > 0)
+            and (MessageDlg('Вы точно хотите удалить строку?',mtConfirmation, [mbYes, mbNo], 0) = mrYes) then
+           begin
+             v_id := DBGrid1.DataSource.DataSet.FieldByName('id').AsString;
+             with DM_main.QryEmpty do
+               begin
+                 SQL.Clear;
+                 SQL.Add('DELETE FROM positions ');
+                 SQL.Add(' WHERE id = ' + v_id + ' ');
+                 try ExecSQL except
+                   on E: EDatabaseError do
+                   ShowMessage (E.Message);
+                 end;
                end;
+             DBGrid1.DataSource.DataSet.Refresh;
              end;
-           DBGrid1.DataSource.DataSet.Refresh;
-         end;
-       end;
+        end;
 // ***** Справочник - Отделы  *****
-  if Pagecontrol1.TabIndex = 1 then
+  if (Pagecontrol1.TabIndex = 1) then
    begin
-       if DBGrid2.DataSource.DataSet.RecordCount = 0 then
+       if DBGrid2.DataSource.DataSet.RecordCount <= 0 then
          ShowMessage('Ничего не выбрано');
-       if DBGrid2.DataSource.DataSet.RecordCount > 0 then
+       if (DBGrid2.DataSource.DataSet.RecordCount > 0)
+           and (MessageDlg('Вы точно хотите удалить строку?',mtConfirmation, [mbYes, mbNo], 0) = mrYes) then
         begin
           v_id := DBGrid2.DataSource.DataSet.FieldByName('id').AsString;
           with DM_main.QryEmpty do
@@ -139,10 +161,35 @@ begin
            DBGrid2.DataSource.DataSet.Refresh;
          end;
        end;
+
 // ***** Справочник - Сотрудники  *****
-  if Pagecontrol1.TabIndex = 2 then
-    showmessage('2');
-  end;
+  if (Pagecontrol1.TabIndex = 2) then
+    begin
+       if DBGrid3.DataSource.DataSet.RecordCount <= 0 then
+         ShowMessage('Ничего не выбрано')
+       else
+         v_fired_flag := DBGrid3.DataSource.DataSet.FieldByName('id_flag_fired').AsInteger;
+       if v_fired_flag = 0  Then
+          ShowMessage('Сотрудник уже отмечен на удаление');
+       if (DBGrid3.DataSource.DataSet.RecordCount > 0)  and (v_fired_flag = 1)
+           and (MessageDlg('Вы точно хотите отметить сотрудника на увольнение?',mtConfirmation, [mbYes, mbNo], 0) = mrYes) then
+        begin
+          v_id := DBGrid3.DataSource.DataSet.FieldByName('id').AsString;
+          v_full_name :=  DBGrid3.DataSource.DataSet.FieldByName('full_name').AsString;
+          with DM_main.QryEmpty do
+            begin
+              SQL.Clear;
+              SQL.Add('Update employees ');
+              SQL.Add('SET flag_fired = 0 ');
+              SQL.Add(' WHERE id = ' + v_id + ' ');
+               try ExecSQL except
+                on E: EDatabaseError do
+                 ShowMessage (E.Message);
+               end;
+             end;
+           DBGrid3.DataSource.DataSet.Refresh;
+         end;
+    end;
 end;
 
 procedure TForm101.bntUpdateClick(Sender: TObject);
@@ -178,7 +225,21 @@ begin
     end;
 // ***** Справочник - Сотрудники  *****
   if Pagecontrol1.TabIndex = 2 then
-    showmessage('2');
+    begin
+    if DBGrid3.DataSource.DataSet.RecordCount = 0 then
+         ShowMessage('Ничего не выбрано');
+      if DBGrid3.DataSource.DataSet.RecordCount > 0 then
+        begin
+          Form106.btnAdd.Visible := False;
+          Form106.btnUpdate.Visible := True;
+          Form106.Edit1.Text := DBGrid3.DataSource.DataSet.FieldByName('full_name').AsString;
+          Form106.MaskEdit1.Text := DBGrid3.DataSource.DataSet.FieldByName('phone').AsString;
+          Form106.DateTimePicker1.Date := StrToDate(DBGrid3.DataSource.DataSet.FieldByName('birthday').AsString);
+          Form106.Combobox2.Text := DBGrid3.DataSource.DataSet.FieldByName('name_dept').AsString;
+          Form106.Combobox1.Text := DBGrid3.DataSource.DataSet.FieldByName('name_pos').AsString;
+          Form106.Show;
+        end;
+    end;
 end;
 
 procedure TForm101.FormCreate(Sender: TObject);
@@ -187,6 +248,7 @@ begin
   DM_main.qryPost.Active := True;
   DM_main.qryDept.MacroByName('name_filter').Active := False;
   DM_main.qryDept.Active := True;
+  DM_main.qryEmp.Active := True;
 end;
 
 procedure TForm101.btnSearchDeptClick(Sender: TObject);
@@ -200,6 +262,62 @@ begin
   else begin
      DM_main.qryDept.MacroByName('name_filter').Active := False;
   end;
+  DM_main.qryDept.Active := True;
+end;
+
+procedure TForm101.TabSheet1Show(Sender: TObject);
+begin
+  btnDelete.Width := 75;
+  btnDelete.Caption := 'Удалить';
+end;
+
+procedure TForm101.TabSheet2Show(Sender: TObject);
+begin
+  btnDelete.Width := 75;
+  btnDelete.Caption := 'Удалить';
+end;
+
+procedure TForm101.TabSheet3Show(Sender: TObject);
+begin
+  btnDelete.Width := 145;
+  btnDelete.Caption := 'Отметка на увольнение';
+end;
+
+procedure TForm101.btnSearchEmpClick(Sender: TObject);
+begin
+  DM_main.qryEmp.Active := False;
+  if Edit2.Text <> '' then
+    begin
+      DM_main.qryEmp.MacroByName('fio_filter').Active := True;
+      DM_main.qryEmp.Params.ParamByName('p_fio').AsString := '%' + Edit2.Text + '%';
+    end
+  else begin
+     DM_main.qryEmp.MacroByName('fio_filter').Active := False;
+  end;
+  DM_main.qryEmp.Active := True;
+end;
+
+procedure TForm101.btnResetPostClick(Sender: TObject);
+begin
+  Edit1.Clear;
+  DM_main.qryPost.Active := False;
+  DM_main.qryPost.MacroByName('name_filter').Active := False;
+  DM_main.qryPost.Active := True;
+end;
+
+procedure TForm101.btnResetEmpClick(Sender: TObject);
+begin
+  Edit2.Clear;
+  DM_main.qryEmp.Active := False;
+  DM_main.qryEmp.MacroByName('fio_filter').Active := False;
+  DM_main.qryEmp.Active := True;
+end;
+
+procedure TForm101.btnResetDeptClick(Sender: TObject);
+begin
+  Edit3.Clear;
+  DM_main.qryDept.Active := False;
+  DM_main.qryDept.MacroByName('name_filter').Active := False;
   DM_main.qryDept.Active := True;
 end;
